split broker in two
